pyproject.toml
[project]
name = "walter-reed-cardiology-agent"
version = "0.1.0"
description = "A2A compliant cardiology referral agent for Dr. Walter Reed's clinic"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server,sqlite]>=0.3.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]

[tool.setuptools]
py-modules = []
